+++ tests/QueryBuilderTest.php
@@ -446,12 +446,19 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
     public function testUpdate(
         string $table,
         array $columns,
-        array|string $condition,
+        array|ExpressionInterface|string $condition,
+        array|ExpressionInterface|string|null $from,
         array $params,
         string $expectedSql,
         array $expectedParams = [],
     ): void {
-        parent::testUpdate($table, $columns, $condition, $params, $expectedSql, $expectedParams);
+        if ($from !== null) {
+            $this->expectException(NotSupportedException::class);
+            $this->expectExceptionMessage(
+                'Yiisoft\Db\Mssql\DMLQueryBuilder::update() does not support UPDATE with FROM clause with SQL Server.',
+            );
+        }
+        parent::testUpdate($table, $columns, $condition, $from, $params, $expectedSql, $expectedParams);
     }
 
     #[DataProviderExternal(QueryBuilderProvider::class, 'upsert')]