+++ CHANGELOG.md
@@ -58,6 +58,7 @@
   and `ShortestBuilder` classes (@Tigrov)
 - Enh #393: Refactor `DMLQueryBuilder::upsert()` method (@Tigrov)
 - Chg #398: Update expression namespaces according to changes in `yiisoft/db` package (@Tigrov)
+- Enh #392: Update `DMLQueryBuilder::update()` method to adapt changes in `yiisoft/db` (@rustamwin)
 
 ## 1.2.0 March 21, 2024